+++ packages/stream_video/lib/src/webrtc/media/constraints/mirror_mode.dart
@@ -1,0 +1,5 @@
+enum MirrorMode {
+  on,
+  off,
+  defaultMode,
+}