+++ packages/stream_video/CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+✅ Added
+
+* Added `mirror` parameter to `CameraConstraints`, which determines whether the camera for the given track should be mirrored or not. When set as null the default is true when `facingMode` is set to `user` and false when facingMode is set to `environment`.
+
 ## 0.3.6
 ✅ Added